created radio group as seperate component
